change demo data

+++ service/app/Services/StatusServiceInterface.php
@@ -16,7 +16,7 @@ interface StatusServiceInterface
      * Получить количество транзакций в секунду
      * @return int
      */
-    public function getTransactionsPerSecond(): int;
+    public function getTransactionsPerSecond(): float;
 
     /**
      * Получить среднее время обработки блока в секундах
